+++ src/audio/gain.h
@@ -8,8 +8,8 @@ static const int INTERNAL_GAIN_MAX = 255;
 
 typedef struct
 {
-    float l;
-    float r;
+    double l;
+    double r;
 } stereo_frame_t;
 
 void set_master_gain(int gain);
@@ -18,6 +18,6 @@ void module_gain_goes_to(int module_gain_maximum);
 
 int get_internal_gain(int module_gain);
 
-stereo_frame_t apply_gain(float pcm, voice_t *voice);
+stereo_frame_t apply_gain(double pcm, voice_t *voice);
 
 #endif //ARCTRACKER_GAIN_H
