+++ src/pcm/mu_law.h
@@ -5,8 +5,8 @@
 
 void precalculate_mu_law();
 
-float *convert_mu_law_to_linear_pcm(const __uint8_t *mu_law_encoded, int no_samples);
+double *convert_mu_law_to_linear_pcm(const __uint8_t *mu_law_encoded, int no_samples);
 
-float convert_to_linear_gain(int logarithmic_gain);
+double convert_to_linear_gain(int logarithmic_gain);
 
 #endif // ARCTRACKER_MU_LAW_H
